US-004/CA-29 - Delete unknown theme

+++ src/themes/default-theme-service.ts
@@ -59,6 +59,12 @@ export class DefaultThemeService implements ThemeService {
     }
 
     deleteTheme(id: string): void {
+        let theme = this.themeRepository.getById(id);
+
+        if (theme === undefined) {
+            throw new ThemeNotFoundError();
+        }
+
         this.themeRepository.delete(id);
     }
